Add files via upload
On the graph on the left, the medium bubble size indicates the relative importance of the topic to State of the Union addresses, demonstrating this topic as the second most relevant. The top three topics spaced out from each other indicates little to no overlap in the topics. The bar chart demonstrates frequency of words for the topic (red bars) in relation to the overall frequency of use for State of the Union addresses (blue bar). Frequent use of words such as program, job, budget, worker, spend, and deficit indicate this topic is related to the U.S. economy.