Add files via upload
On the graph on the left, the large bubble size indicates the importance of the topic to State of the Union addresses. The top three topics spaced out from each other indicates little to no overlap in the topics. The bar chart demonstrates frequency of words for the topic (red bars) in relation to the overall frequency of use for State of the Union addresses (blue bar). Frequent use of words such as interstate, tariff, farmer, railroad, canal, and veteran indicate this topic is related to domestic spending/government programming.